move PyDocument to global package
