+++ python/src/lib.rs
@@ -18,8 +18,9 @@ use raphtory_core::python::{
         algorithms::*,
         graph_gen::*,
         graph_loader::*,
-        vectors::{generate_property_list, PyDocument, PyVectorisedGraph},
+        vectors::{generate_property_list, PyVectorisedGraph},
     },
+    types::wrappers::document::PyDocument,
 };
 
 macro_rules! add_functions {
