+++ raphtory/src/python/types/wrappers/mod.rs
@@ -1,2 +1,3 @@
+pub mod document;
 pub(crate) mod iterators;
 pub(crate) mod prop;
